Apply changes to SRFI 113 impl. from John Cowan.
I have one issue that should not affect finalization, because it only changes the implementation. I would like to
change the SRFI 113 implementation to make use of hashmap-based sets rather than ordered sets. This makes
it compatible with the sample implementation in the SRFI 113 repository. Changing all instances of "mapping" to
"hashmap" in srfi/113.scm, defining hashmap-fold/reverse as hashmap-fold in the same file, and changing the
import of (srfi 146) to (srfi 146 hash) in srfi/113.sld is ALMOST enough.

However, there is a bad recursion. srfi/146/hash.scm depends on gleckler/hamt-map.sld, which exports a trivial
procedure called 'make-string-set', which invokes the 'set' procedure from srfi/113.scm. Since with the above
change, srfi/113.scm depends on srfi/146/hash.scm, Chibi cannot load either library.

This problem is resolved by removing make-string-set from its current location and adding it to
gleckler/hamt-map-test.scm, which is the only place it is actually used.

Unfortunately, this makes the tests for SRFI 113 crash mysteriously. I have attached a zip of the changed files
(diffs seem to get corrupted).